Stop using deprecated function AServiceManager_getService()
AServiceManager_getService() is marked as deprecated and causes compiler
warnings when building vold.  It is deprecated because it waits for up
to 5 seconds for the service to become available, which is an arbitrary
amount.  Callers either want no wait (AServiceManager_checkService()) or
an indefinite wait (AServiceManager_waitForService()).  vold wants an
indefinite wait, so switch to using AServiceManager_waitForService().

Test: 'mmm system/vold' and verified that the warnings are gone
Test: Boot Cuttlefish
Flag: EXEMPT bugfix
Change-Id: Id9e9686f6d4ab8ef3beb23f4a85efb850dc8fece